Move edit mode drag preview controller functions to a new hook
